fix: fixed problem concerning how 'setInterval' can only be cleared the first time, subsequent attempts to clear 'setInterval' invocations seem to have no effect. The problem was due to storing the objecct id of 'setInterval' in reactive state using the 'useState' hook and then clearing the object stored as state with 'clearInterval'.
A thought about why this is the case is that unlike the function returned from 'useState' that is used for updating state, the identity of the state itself doesn't persist through renders (see https://reactjs.org/docs/hooks-reference.html#usestate), as each time a function component is rendered, its hooks are called one by one in the order in which they're defined. And a function component re-renders each time its state has changed since its last render. With this in mind, given how the same App component is keeping track of the board as state and the nature of the game of life involves multiple updates to the board in a narrow time frame, it is likely that once the game has commenced, these changes to the board's state will trigger component re-renders, which will in turn interfere with the state variable that's holding the 'setInterval' id of the currently executing 'setInterval'. Hence, I think that when attempting to then clear the interval stored in the state variable, react struggles to keep track of the correct interval that was set as the new state, and so we fail to clear 'setInterval' invocations that come after the initial one. (Although we could perhaps try resetting the state variable when we invoke 'clearInterval'.)

The workaround currently adopted is to move the variable that holds a reference to the first function argument of 'setInterval' outside of the react component, since variables defined outside of the react component maintain a stable identity through time or through component renders. This way, we handle this ourselves rather than make use of the 'useState' hook.